### 2024-02-25 #6409 @editorialbot
[PRE REVIEW]: JasPer: A Portable Flexible Software Toolkit for Image Coding/Processing
**Submitting author:** @mdadams (<a href="http://orcid.org/0000-0001-6163-7062">Michael D. Adams</a>)
**Repository:** https://github.com/jasper-software/jasper
**Branch with paper.md** (empty if default branch): mdadams-joss
**Version:** 4.2.1
