### 2024-02-25 #6409 @danielskatz
👋 @adamltyson - would you be willing to edit this JOSS submission?
