### 2024-02-25 9784a72
Removed an unused entry from the BibTeX file for the JOSS paper.
