+++ joss_paper/paper.bib
@@ -247,18 +247,6 @@
 }
 
 @inproceedings{
-	Wu_2015a,
-	author = "Wu, Yi-Ting and Wu, Hung-Tsai and Chang, Wen-Whei",
-	booktitle = "2015 12th International Joint Conference on e-Business and Telecommunications (ICETE)", 
-	title = "Compressed domain ECG biometric identification using {JPEG}2000", 
-	year = 2015,
-	volume = 05,
-	pages = "5--13",
-	isbn = "978-9-8975-8140-3",
-	url = "https://ieeexplore.ieee.org/document/7518100",
-}
-
-@inproceedings{
 	Verhack_2015a,
 	author = "Verhack, Ruben and Lange, Lieven and Lambert, Peter and Van de Walle, Rik and Sikora, Thomas",
 	booktitle = "2015 Picture Coding Symposium (PCS)", 
