### 2024-02-25 #6409 @mdadams
@danielskatz As far as I can tell, the DOI suggested by editorialbot was for an entry in my BibTeX file that was not used in the submitted version of my paper.  I had forgotten to delete this unused entry, but I have just removed it now.  I will try the editorialbot commands that you suggested in a separate comment to see if editorialbot is happier with the revised document (with the unused BibTeX entry deleted).
