### 2024-02-26 #6409 @adamltyson
> 👋 @adamltyson - would you be willing to edit this JOSS submission?

Hi @danielskatz, yes, happy to.
