### 2024-02-28 #6409 @mdadams
@adamltyson I do not have any suggestions for specific individuals to serve as reviewers.  I would assume that anyone with a good knowledge of the C language and some basic image processing background should be able to perform the review.  There appears to be quite a few people in the JOSS reviewer list who are identified as knowing C and having topic areas that include image processing.  So, presumably, any of these people should be fine to serve as reviewers.  Does that help at all?  Since this is my first time submitting to JOSS, I suspect that I probably could not make a better suggestion for potential reviewers than whoever you would choose yourself.
