### 2024-02-29 #6409 @adamltyson
:wave: @tomelse,  @justusschock, @vitorsr, would you be able to review this submission to JOSS? I think you've all reviewed for JOSS before, but if you have any questions about the process, please let me know. Thanks!
