### 2024-02-29 #6409 @vitorsr
@adamltyson I think it would be best if we could somehow invite a contributor or maintainer of @uclouvain’s [OpenJPEG](https://github.com/uclouvain/openjpeg) project. Their expertise would be most valuable in this review.
