### 2024-02-29 #6409 @adamltyson
> I think it would be best if we could somehow invite a contributor or maintainer of https://github.com/uclouvain’s [OpenJPEG](https://github.com/uclouvain/openjpeg) project. Their expertise would be most valuable in this review.

Thanks @vitorsr. That would be useful, but in the interest of time I think it's best to move on with the reviewers we have.
