### 2024-03-11 6f61244
Updated the NEWS document and bumped the version number.
