+++ NEWS.txt
@@ -1,3 +1,8 @@
+4.2.2 (2024-03-11)
+==================
+
+* Fix minor build issue (#374).
+
 4.2.1 (2024-02-18)
 ==================
 
