### 2024-03-18 #6431 @justusschock
Hey @adamltyson , the [license](https://github.com/jasper-software/jasper/blob/master/LICENSE.txt) does seem to be a non-standard osi-license. It looks like some sort of extended MIT to me. Could you clarify if the license still works?
