### 2024-03-18 #6431 @adamltyson
> Hey @adamltyson , the [license](https://github.com/jasper-software/jasper/blob/master/LICENSE.txt) does seem to be a non-standard osi-license. It looks like some sort of extended MIT to me. Could you clarify if the license still works?

Thanks for raising this @justusschock.

@mdadams as per the [journal guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#software-license). Would you be able to relicense the code under an [OSI approved license](https://opensource.org/license)? If it is an existing, approved license, could you let us know? I had a look and couldn't find this specific license anywhere.
