### 2024-03-28 #6431 @mdadams
@adamltyson:

While I appreciate that the license for JasPer (formally known as the "JasPer 2.0" license) has not been submitted for approval by the OSI, it has been approved by a number of very well-respected organizations in the open-source community, including both the Fedora Project and OpenSUSE Project.  You can confirm for yourself that the JasPer 2.0 license is listed as one of the licenses that has been vetted by Fedora Legal and approved as a valid open-source license for use by the Fedora Project at:

  <https://docs.fedoraproject.org/en-US/legal/allowed-licenses>

(Refer to the table titled "Allowed Licenses".)  You can also confirm that the JasPer 2.0 license has been approved by the OpenSUSE Project for use in OpenSUSE Linux at:

  <https://en.opensuse.org/openSUSE:Accepted_licences>

(Refer to the table under the heading "Good Licenses'.)  In case this might also be helpful information, the table in which the JasPer license is listed in this (OpenSUSE) web page indicates that this license is FSF Free, GPLv2 compatible, and GPLv3 compatible.  JasPer is included in many other Linux and BSD Unix distributions, Homebrew, and Conan, as well as many other projects.  Most, if not all, of these projects require the software they include to employ an open-source license.  The JasPer 2.0 license was inspired by the MIT and BSD licenses.  Essentially, the JasPer 2.0 license is a MIT license with wording added to emphasize the disclaimer and prevent the names of the copyright holders from being used to promote products based on JasPer without prior permission.

With respect to OSI approval, there are only two options for making JasPer use a license that is formally approved by the OSI:

1. JasPer would need to be released under a different license that has already been approved by the OSI.

2. The existing JasPer license would need to be submitted to the OSI for approval as a legacy license and subsequently approved by the OSI.

Option 1 (i.e., changing the JasPer license) is not feasible.  Since JasPer has a history of almost 25 years, changing the license for this software is simply not possible.  Changing the license would require obtaining the consent of the original copyright holders as well as all of the very many individuals who have contributed bug fixes, patches, and other changes over a period of almost 25 years.  Just identifying all of these individuals/parties would be problematic.

Option 2 (i.e., OSI approval of the current license) also appears not to be feasible.  The OSI web site states that legacy licenses will only be considered for approval if they are used by more than twenty projects maintained by different unrelated entities.  Since the JasPer 2.0 license is only used by JasPer, it does not meet this criterion for being considered.

While I appreciate that it is extremely important to JOSS that it only publish software that employs a license that is truly open source, it would seem to be somewhat contrary to the spirit of the journal not to consider a software project that clearly meets the criterion of being open source, except for the technicality that its license has not been approved by the OSI.  (Note that the OSI was founded in 1998, which is only one year prior to the start of the development of JasPer.)  With all of the above said, however, I do respect that the decision of whether JasPer is a valid submission to JOSS is not mine to make.  Clearly, this decision is solely at the discretion of the journal.  I would only ask that you please consider the somewhat special circumstances surrounding JasPer when making your final decision.  As I have demonstrated in the paper accompanying the submission, JasPer has helped the research community over many years, and JasPer is very clearly used in many projects that specifically require an open-source license.
