### 2024-03-29 6d084c5
Fixes #381.
Added a missing check to the jpc_dec_process_sod function of the JPC codec.
Added another image to the test set.
