### 2024-03-30 2fe7d2d
Bumped the version number and updated the NEWS document.
