+++ NEWS.txt
@@ -1,3 +1,8 @@
+4.2.3 (2024-03-30)
+==================
+
+* Added a missing check in the JPC codec (#381).
+
 4.2.2 (2024-03-11)
 ==================
 
