### 2024-04-04 #6431 @danielskatz
@mdadams - as JOSS is an affiliate of OSI, we have a firm policy that we will only review software that has an OSI-approved license. Given this, and your discussion above that says that this software is not and will not have an OSI-approved license, it seems like we should end the review process and mark this as withdrawn. Please let me know if you agree.

If so, I will point you to https://www.software.ac.uk/top-tip/which-journals-should-i-publish-my-software for some other options, as well as https://joss.readthedocs.io/en/latest/submitting.html#other-venues-for-reviewing-and-publishing-software-packages for other suggestions for how you might receive credit for your work.  Of course, none of this is a statement about the quality or utility of your software, just that it does not fit the JOSS requirements.
