### 2024-04-10 29154a5
Updated the NEWS document (to add a reference to a CVE).
