+++ NEWS.txt
@@ -1,7 +1,7 @@
 4.2.3 (2024-03-30)
 ==================
 
-* Added a missing check in the JPC codec (#381).
+* Added a missing check in the JPC codec (#381) (CVE-2024-31744).
 
 4.2.2 (2024-03-11)
 ==================
