### 2024-04-27 2d6b6e8
Fixes #383.
Added some missing checks to the jas_heic_decode function in the
HEIC codec.
Added support for handling HEIC images in the run_test_1 script.
Added a HEIC image to the test set.
