### 2024-04-27 8766848
Updated the NEWS document and bumped the version number.
