+++ NEWS.txt
@@ -1,3 +1,9 @@
+4.2.4 (2024-04-27)
+==================
+
+* Added some missing checks to the jas_heic_decode function in the
+  HEIC codec (#383).
+
 4.2.3 (2024-03-30)
 ==================
 
