### 2024-05-10 #6431 @danielskatz
@mdadams - We will proceed to reject this in 2 weeks unless we hear from you about the software being changed to an OSI-approved license.
